[Fix] #17 - SettingBaseViewController 사용하여 수정
